Refactor proposals list route
