+++ src/features/organizer/event/Event.jsx
@@ -3,6 +3,7 @@ import { Route, Routes, useParams } from 'react-router-dom'
 
 import EventPage from 'features/event'
 import EventEdit from './edit'
+import EventProposals from './proposals'
 
 const Event = () => {
   const { eventId } = useParams()
@@ -10,6 +11,7 @@ const Event = () => {
     <Routes>
       <Route path="/" element={<EventPage eventId={eventId} />} />
       <Route path="/edit/*" element={<EventEdit eventId={eventId} />} />
+      <Route path="/proposals" element={<EventProposals eventId={eventId} />} />
     </Routes>
   )
 }
