+++ src/features/organizer/event/proposals/proposals.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import { useLocation } from 'react-router-dom'
 
 import HasRole from 'screens/components/hasRole'
 import { ROLE_OWNER_OR_MEMBER } from 'firebase/constants'
@@ -14,14 +15,25 @@ import ProposalsPaging from './proposalsPaging'
 const Proposals = ({ eventId }) => {
   const { user } = useAuth()
 
+  const { search } = useLocation()
+  const params = new URLSearchParams(search)
+  const filters = {
+    search: params.get('search'),
+    state: params.get('state'),
+    ratings: params.get('ratings'),
+    formats: params.get('formats'),
+    categories: params.get('categories'),
+    sortOrder: params.get('sortOrder'),
+  }
+
   return (
     <div>
       <ProposalsHeader eventId={eventId} />
-      <ProposalsFilters userId={user.uid} />
+      <ProposalsFilters eventId={eventId} />
       <HasRole of={ROLE_OWNER_OR_MEMBER} forEventId={eventId}>
         <ProposalsToolbar eventId={eventId} userId={user.uid} />
       </HasRole>
-      <ProposalsList eventId={eventId} userId={user.uid} />
+      <ProposalsList eventId={eventId} userId={user.uid} filters={filters} />
       <ProposalsPaging />
     </div>
   )
