+++ src/features/organizer/event/proposals/proposals.listeners.js
@@ -3,7 +3,6 @@ import { when } from 'k-ramel'
 import * as reactions from './proposals.reactions'
 
 export default [
-  when('@@krml/LISTENERS>ADDED>PROPOSALS')(reactions.setProposalFiltersFromRouter),
   when('@@ui/ON_LOAD_EVENT_PROPOSALS')(reactions.loadProposals),
   when('@@ui/ON_SELECT_PROPOSAL')(reactions.selectProposal),
   when('@@ui/SELECT_ALL_PROPOSALS')(reactions.selectAllProposal),
@@ -11,5 +10,4 @@ export default [
   when('@@ui/ACCEPT_PROPOSALS')(reactions.acceptProposals),
   when('@@ui/REJECT_PROPOSALS')(reactions.rejectProposals),
   when('@@ui/ON_ADD_PROPOSAL_TO_SELECTION')(reactions.addProposalToSelection),
-  when('@@ui/CHANGE_PROPOSAL_FILTER')(reactions.changeFilter),
 ]
