+++ src/features/organizer/event/proposals/proposalsHeader/index.js
