+++ src/features/organizer/event/proposals/proposalsList/index.js
