+++ src/features/organizer/event/proposals/proposalsList/proposalInfo/index.js
