+++ src/features/organizer/event/proposals/proposalsList/proposalInfo/proposalInfo.css
