+++ src/features/organizer/event/proposals/proposalsList/proposalInfo/proposalInfo.jsx
