+++ src/features/organizer/event/proposals/proposalsList/proposalSubtitle/index.js
