+++ src/features/organizer/event/proposals/proposalsList/proposalSubtitle/proposalSubtitle.jsx
