+++ src/features/organizer/event/proposals/proposalsList/proposalsList.jsx
@@ -1,0 +1,71 @@
+import React, { useEffect } from 'react'
+import PropTypes from 'prop-types'
+
+import { withSizes } from 'styles/utils'
+import { List, ListItem } from 'components/list'
+import ProposalSubtitle from './proposalSubtitle'
+import ProposalInfo from './proposalInfo'
+import './proposalsList.css'
+
+const Proposals = ({
+  eventId,
+  proposals,
+  proposalsSelection,
+  deliberationActive,
+  blindRating,
+  onSelect,
+  onAddProposalToSelection,
+  onLoad,
+  filters,
+  isMobile,
+}) => {
+  useEffect(() => {
+    onLoad({ filters })
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [filters])
+
+  return (
+    <List
+      className="event-proposals"
+      array={proposals}
+      renderRow={(proposal) => (
+        <ListItem
+          key={proposal.id}
+          title={proposal.title}
+          subtitle={
+            !isMobile && (
+              <ProposalSubtitle eventId={eventId} proposal={proposal} blindRating={blindRating} />
+            )
+          }
+          info={<ProposalInfo eventId={eventId} proposal={proposal} isMobile={isMobile} />}
+          onSelect={() => onSelect(proposal.id)}
+          onCheckboxChange={() => onAddProposalToSelection(proposal.id)}
+          checked={!!proposalsSelection.includes(proposal.id)}
+          checkboxDisabled={!deliberationActive}
+        />
+      )}
+    />
+  )
+}
+
+Proposals.propTypes = {
+  eventId: PropTypes.string.isRequired,
+  proposals: PropTypes.arrayOf(PropTypes.object),
+  proposalsSelection: PropTypes.arrayOf(PropTypes.string),
+  deliberationActive: PropTypes.bool,
+  blindRating: PropTypes.bool,
+  onSelect: PropTypes.func.isRequired,
+  onAddProposalToSelection: PropTypes.func.isRequired,
+  isMobile: PropTypes.bool.isRequired,
+  onLoad: PropTypes.func.isRequired,
+  filters: PropTypes.object.isRequired,
+}
+
+Proposals.defaultProps = {
+  proposals: [],
+  proposalsSelection: [],
+  deliberationActive: false,
+  blindRating: false,
+}
+
+export default withSizes(Proposals)
