+++ src/features/organizer/event/proposals/proposalsPaging/proposalsPaging.module.css
